pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-health-coach"
version = "0.1.0"
description = "AI-powered health coaching platform. Generates personalized health insights from wearable biometric data using local LLMs via LiteLLM, with dbt transformations on PostgreSQL and Evidence.dev dashboards."
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Nicholas Hidalgo", email = "analytics@nicholashidalgo.com"}]
readme = "README.md"

dependencies = [
    "litellm>=1.40.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.30",
    "pydantic>=2.7.0",
    "typer[all]>=0.12.0",
    "structlog>=24.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "faker>=25.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
coach = "ai_health_coach.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[tool.hatch.build.targets.wheel]
packages = ["src"]
